refactor: split .env files to each app
