+++ backend/apps/notifications/.env.example
@@ -1,0 +1,25 @@
+NODE_ENV=production
+
+REDIS_HOST=redis
+REDIS_PORT=6379
+
+REDIS_HOST_EXTERNAL=localhost
+REDIS_PORT_EXTERNAL=6379
+
+POSTGRES_USER=db-user
+POSTGRES_PASSWORD=db-pass
+POSTGRES_DB=weather-api
+
+DB_HOST=db
+DB_PORT=5432
+DATABASE_URL=postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${DB_HOST}:${DB_PORT}/${POSTGRES_DB}
+
+DB_HOST_EXTERNAL=localhost
+DB_PORT_EXTERNAL=5432
+DATABASE_URL_EXTERNAL=postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${DB_HOST_EXTERNAL}:${DB_PORT_EXTERNAL}/${POSTGRES_DB}
+
+SENDGRID_API_KEY=
+SENDGRID_FROM_EMAIL=
+
+LOG_LEVEL="info"
+LOG_FILE_PATH=log/app.log
