+++ backend/apps/notifications/.env.test
@@ -1,29 +1,25 @@
-NODE_ENV=production
-API_PORT=8000
-API_URL=http://localhost:8000/api
+NODE_ENV=test
 
-REDIS_HOST=redis
+REDIS_HOST=redis-test
 REDIS_PORT=6379
 
 REDIS_HOST_EXTERNAL=localhost
-REDIS_PORT_EXTERNAL=6379
+REDIS_PORT_EXTERNAL=6381
 
-POSTGRES_USER=db-user
-POSTGRES_PASSWORD=db-pass
-POSTGRES_DB=weather-api
+POSTGRES_USER=db-test
+POSTGRES_PASSWORD=test-pass
+POSTGRES_DB=weather-api-test
 
-DB_HOST=db
+DB_HOST=db-test
 DB_PORT=5432
 DATABASE_URL=postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${DB_HOST}:${DB_PORT}/${POSTGRES_DB}
 
 DB_HOST_EXTERNAL=localhost
-DB_PORT_EXTERNAL=5432
+DB_PORT_EXTERNAL=5434
 DATABASE_URL_EXTERNAL=postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${DB_HOST_EXTERNAL}:${DB_PORT_EXTERNAL}/${POSTGRES_DB}
 
-SENDGRID_API_KEY=
-SENDGRID_FROM_EMAIL=
-
-WEATHER_API_KEY=
+SENDGRID_API_KEY=SG.dummy
+SENDGRID_FROM_EMAIL=test@gmail.com
 
 LOG_LEVEL="info"
-LOG_FILE_PATH=log/app.log
+LOG_FILE_PATH=log/app-test.log
