+++ backend/apps/notifications/tsconfig.test.json
@@ -1,5 +1,5 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["src/**/*"],
+  "include": ["src/**/*.test.ts"],
   "exclude": ["node_modules"]
 }
