+++ backend/apps/weather/.env.test
@@ -20,9 +20,6 @@ DB_HOST_EXTERNAL=localhost
 DB_PORT_EXTERNAL=5434
 DATABASE_URL_EXTERNAL=postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${DB_HOST_EXTERNAL}:${DB_PORT_EXTERNAL}/${POSTGRES_DB}
 
-SENDGRID_API_KEY=SG.dummy
-SENDGRID_FROM_EMAIL=test@gmail.com
-
 WEATHER_API_KEY=test_key
 
 LOG_LEVEL="info"
