+++ backend/apps/weather/package.json
@@ -6,17 +6,17 @@
   "scripts": {
     "build": "tspc",
     "start": "node dist/index.js",
-    "dev": "npm run db:generate && dotenv -e ../../.env.development -- nodemon --watch 'src/**/*.ts' --exec 'ts-node -r tsconfig-paths/register' src/index.ts",
+    "dev": "npm run db:generate && dotenv -e .env.development -- nodemon --watch 'src/**/*.ts' --exec 'ts-node -r tsconfig-paths/register' src/index.ts",
     "lint": "eslint 'src/**/*.ts'",
     "lint:fix": "eslint 'src/**/*.ts' --fix",
     "type-check": "tsc --noEmit",
     "test": "npm run test:unit && npm run test:integration",
     "test:unit": "cross-env NODE_ENV=test jest --config jest.unit.config.js --silent --runInBand",
-    "test:integration:local": "dotenv -e ../../.env.test -- sh -c 'npm run db:migrate-test && jest --config jest.integ.config.js --runInBand --silent'",
+    "test:integration:local": "dotenv -e .env.test -- sh -c 'npm run db:migrate-test && jest --config jest.integ.config.js --runInBand --silent'",
     "test:integration": "bash ./scripts/run-integration-tests.sh",
     "test:arch": "dependency-cruise --config .dependency-cruiser.js src",
     "db:generate": "prisma generate",
-    "db:migrate-test": "dotenv -e ../../.env.test -- sh -c 'DATABASE_URL=\"$DATABASE_URL_EXTERNAL\" npx prisma migrate reset --force'"
+    "db:migrate-test": "dotenv -e .env.test -- sh -c 'DATABASE_URL=\"$DATABASE_URL_EXTERNAL\" npx prisma migrate reset --force'"
   },
   "prisma": {
     "seed": "ts-node ../../prisma/seed.ts",
