+++ backend/apps/weather/src/infrastructure/queue/jobs/update-weather-data/processor.ts
@@ -7,7 +7,6 @@ import {
   UpdateWeatherDataJobData,
   WeatherData,
 } from "@common/generated/proto/job_pb";
-import { env } from "@/config/env";
 import { IQueueService } from "@common/shared/ports";
 
 export class UpdateWeatherDataProcessor {
@@ -35,7 +34,7 @@ export class UpdateWeatherDataProcessor {
       const { email, city, unsubscribeToken, id } = subscription;
 
       const weatherData = await this.weatherProvider.getWeatherData(city);
-      const unsubscribeUrl = `${env.API_URL}/api/unsubscribe/${unsubscribeToken}`;
+      const unsubscribeUrl = `/api/unsubscribe/${unsubscribeToken}`;
 
       const emailJobData = new SendWeatherUpdateEmailJobData({
         email,
