+++ backend/docker-compose.yml
@@ -4,7 +4,7 @@ services:
     image: postgres:17.5
     container_name: db
     env_file:
-      - .env
+      - apps/weather/.env
     volumes:
       - ./db-data/prod:/var/lib/postgresql/data
     ports:
@@ -22,7 +22,7 @@ services:
     image: redis:8.0
     container_name: redis
     env_file:
-      - .env
+      - apps/weather/.env
     ports:
       - 6379:6379
     networks:
@@ -49,7 +49,7 @@ services:
   weather-service:
     container_name: weather-service
     env_file:
-      - .env
+      - apps/weather/.env
     ports:
       - 8000:8000
     networks:
@@ -67,7 +67,7 @@ services:
   notifications-service:
     container_name: notifications-service
     env_file:
-      - .env
+      - apps/notifications/.env
     networks:
       - weather-api-network
     restart: always
